Improved the DeleteObjectInputController. -Stopped camera movement when something is selected. -Preview deleted object.

+++ core/src/com/marama/view/controllers/DeleteObjectInputController.java
@@ -7,20 +7,35 @@ import com.marama.view.screens.GameScreen;
 
 public class DeleteObjectInputController extends InputAdapter {
     private GameScreen gameScreen;
-
+    private SelectableInstance DeletionSelected = null;
     public DeleteObjectInputController(GameScreen gameScreen) {
         this.gameScreen = gameScreen;
     }
 
+    private float deletionOpacity = 0.5f;
+
+    @Override
+    public boolean touchDown(int screenX, int screeny, int pointer, int button) {
+        Ray ray = gameScreen.world.getPerspectiveCamera().getPickRay(screenX, screeny);
+        DeletionSelected = ((SelectableInstance) gameScreen.world.getClosestModelInstance(ray));
+        DeletionSelected.setOpacity(deletionOpacity);
+        return false;
+    }
+
+    @Override
+    public boolean touchDragged(int screenX, int screenY, int pointer) {
+        return DeletionSelected != null;
+    }
     @Override
     public boolean touchUp(int screenX, int screenY, int pointer, int button) {
         Ray ray = gameScreen.world.getPerspectiveCamera().getPickRay(screenX, screenY);
         SelectableInstance instance = (SelectableInstance) gameScreen.world.getClosestModelInstance(ray);
 
-        if (instance != null) {
+        if (instance != null && instance == DeletionSelected) {
             gameScreen.world.deleteObject(instance);
         }
-
+        DeletionSelected.setOpacity(1f);
+        DeletionSelected = null;
         return false; // Continue to the next 'touchDragged' listener.
     }
 }
